+++ react/assets/forus-platform/scss/_common/dashboard.scss
@@ -2311,72 +2311,6 @@ body {
         }
     }
 
-    &.card-block-requests {
-        box-shadow: 0 2px 5px rgba(25, 39, 52, 0.11);
-        border-radius: var(--border-radius);
-        margin: 20px 0;
-        color: #646f79;
-        background-color: #fff;
-        border: 1px solid #d4d9dd;
-
-        .card-section {
-            padding: 10px 15px;
-        }
-
-        .card-heading {
-            flex: 1;
-            margin: 0;
-            align-self: center;
-            font: 600 13px/13px var(--base-font);
-            letter-spacing: -0.2px;
-            color: #404040;
-            cursor: pointer;
-        }
-
-        .button {
-            font: 600 12px/12px var(--base-font);
-            color: #134478;
-        }
-
-        .mdi {
-            color: #134478;
-            margin-right: 6px;
-        }
-
-        &.card-block-requests-danger {
-            color: #404040;
-            background-color: #fff5f0;
-            border: 1px solid #e63b3b;
-
-            .mdi {
-                color: #e63b3b;
-            }
-
-            .button {
-                .mdi {
-                    margin-left: 3px;
-                    color: #134478;
-                }
-            }
-        }
-
-        &.card-block-requests-warning {
-            background-color: #fcfaf6;
-            border: 1px solid #f8e39a;
-            color: #404040;
-
-            .mdi {
-                color: #ffb05a;
-            }
-
-            .button {
-                .mdi {
-                    color: #ffb05a;
-                }
-            }
-        }
-    }
-
     &.card-block-request-record {
         .card-header {
             .card-title {
